--- server/controller/UserController.go
@@ -1,3 +1,0 @@
-package controller
-
-// Will move user data over here. In main for now
