+++ server/models/user.go
@@ -1,8 +1,9 @@
 package models
 
 type User struct {
-	Id       int    `json:"id"`
+	Id       string `json:"id"`
 	Email    string `json:"email"`
 	Username string `json:"username"`
+	Password string `json:"-"`
 	IsActive bool   `json:"isActive"`
 }
